メソッド切り分けと命名修正
